feat: 사용자 추천 인프라 수치 - 점수 반영

+++ Back/src/main/java/com/example/back/dashboard/service/DashboardService.java
@@ -56,7 +56,7 @@ public class DashboardService {
 		return infraCountList.stream().map(infraCount -> new InfraTypeCountDto(
 			infraCount.get("dongName", String.class),
 			infraCount.get("infraTypeName", String.class),
-			infraCount.get("infraTypeScore", Double.class),
+			infraCount.get("infraTypeScore", Integer.class),
 			infraCount.get("totalCount", Long.class)))
 			.collect(Collectors.toList());
 	}
